fix type for price/sale

+++ database/migrations/2022_08_16_055449_create_games_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->text('desc')->nullable();
             $table->text('content')->nullable();
             $table->string('cover')->nullable();
-            $table->integer('price')->nullable();
-            $table->integer('sale')->nullable();
+            $table->float('price')->nullable();
+            $table->float('sale')->nullable();
             $table->integer('category_id')->nullable();
             $table->boolean('xbox_one')->default(true);
             $table->boolean('xbox_sx')->default(false);
